create project-component file
