--- client/pages/edit-project/edit-project.html
@@ -1,37 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>HYF Connect</title>
-    <link href="../../libs/fontawesome/css/all.css" rel="stylesheet" />
-    <link rel="preconnect" href="https://fonts.googleapis.com" />
-      <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-      <link href="../../libs/fontawesome/css/all.css" rel="stylesheet"/>
-      <link
-        href="https://fonts.googleapis.com/css2?family=Montserrat:wght@500;700&family=Poppins:wght@400;600&display=swap"
-        rel="stylesheet"
-      />
-    <link rel="stylesheet" href="../../styles/styles.css"/>
-    <link rel="stylesheet" href="../../styles/add-new-project.css" />
-    <link
-      href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css"
-      rel="stylesheet"
-      integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3"
-      crossorigin="anonymous"
-    />
-    <script
-      src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"
-      integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p"
-      crossorigin="anonymous"
-    ></script>
-
-  </head>
-
-  <body>
-    <script src="https://unpkg.com/vue@3.x"></script>
-    <script src="../../libs/autocomplete.js"></script>
-    <div id="add-new-project"></div>
-    <script type="module" src="./add-new-project.js"></script>
-  </body>
-</html>
